make metric struct re-usable
